backoffice/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pandas
numpy
scikit-learn
python-dotenv
redis
slowapi
